requirements.txt
asgiref==3.7.2
bcrypt==4.0.1
Django==4.2.2
django-cors-headers==4.1.0
django-rest-auth==0.9.5
djangorestframework==3.14.0
PyJWT==2.7.0
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
